third_party/blink/web_tests/platform/generic/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-remove-src-expected.txt

This is a testharness.js-based test.
FAIL invoking resource selection by setting src; await stable state assert_equals: expected 0 but got 3
Harness: the test ran to completion.
